Added login.html and login.js to get the access token
This work. I can see the hash part of the url in my login.js
script. This means that now the CLIENT can submit requests to
the spotify api using the token and we should be all good.

That also means that I think Express doesn't need to do anything
else.... I think...